runtime: split seccomp functionality out of seccomp demo
