+++ runtime/seccomp_filter.h
@@ -1,0 +1,2 @@
+#pragma once
+int configure_seccomp(void);
